refactor(RTE): rename createRTE util file
