+++ packages/components/src/RichTextEditor/utils/commands/removeMark.spec.ts
@@ -1,5 +1,5 @@
 import { findByText, getByText, waitFor } from "@testing-library/dom"
-import { createRichTextEditor } from "../core/create"
+import { createRichTextEditor } from "../core"
 import { simulateSelectionByText } from "./fixtures/helpers"
 import { testEditorStateWithMarks, testSchema } from "./fixtures/test-state"
 import { removeMark } from "./removeMark"
