+++ packages/components/src/RichTextEditor/utils/commands/updateMark.spec.ts
@@ -1,5 +1,5 @@
 import { getByRole, getByText, waitFor } from "@testing-library/dom"
-import { createRichTextEditor } from "../core/create"
+import { createRichTextEditor } from "../core"
 import { simulateSelectionByText } from "./fixtures/helpers"
 import { testEditorStateWithMarks, testSchema } from "./fixtures/test-state"
 import { updateMark } from "./updateMark"
