+++ packages/components/src/RichTextEditor/utils/core/createRichTextEditor.ts
